// src/app/Models/Task.ts
/**
 *
 */
export class Task
{
    id: number;
    user_id: any = '';
    duration: string;
    project_id: any = '';
    category_id: any = '';
    completion_date: string;
    title: string;
    is_active: number;
    api_token: string;
    created_by: number;
    created_at: string;
    updated_at: string;
    updated_by: number;
    status: any = '';
    description: string;
    remarks: string;
}